Updated Website
The main page was renamed to index.html to work with GitHub pages. The mw_styles_txt.html was renamed to mw_styles.html. Styles for the html and body background were added. The body footer was also styled. The address text was formatted to be centered. The navigation bar was styled so that they could be used. The h1 in the article was made larger and the spacing was changed. The h1 and paragraphs within the article was styled. The block quotes were also formatted.